requirements.txt
websocket-client~=1.8.0
pandas~=2.2.3
logging~=0.4.9.6
streamlit~=1.41.1
python-dotenv~=1.0.1
ta==0.11.0  # of vervang '0.10.2' door de geïnstalleerde versie
pyyaml~=6.0.2
plotly~=5.24.1
pytz
schedule==1.1.0
requests~=2.32.3
joblib==1.3.1
python-dotenv
python-bitvavo-api
... (rest)
